-moved the notification to common
